+++ bcl/src/Core/CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGE LOG
 
-## 0.1.0 initial creation
+## 0.1.1 initial creation
 
-- created.
+- Add Symbol struct for js/ruby like symbols of interned strings.
+- Add Fs.ChangeOwner function to (chown) change the owner of a file.
+
+## 0.1.0 initial creation